+++ tools/starcoinutil_test.go
@@ -22,7 +22,7 @@ func TestGetStarcoinNodeHeight(t *testing.T) {
 
 func TestGetTransactionProof(t *testing.T) {
 	restclient := NewRestClient()
-	var eventIndex uint64 = 1
+	var eventIndex int = 1
 	p, err := GetTransactionProof("https://halley-seed.starcoin.org", restclient,
 		"0x815764e45c2f300cfd90e9a693207c0d4f8f2ad1e3e9774f489534f1ab74e3d5", 69937, &eventIndex)
 	if err != nil {
